+++ lib/hexapdf/utils/sorted_tree_node.rb
@@ -174,6 +174,7 @@ module HexaPDF
           elsif node.key?(:Kids)
             index = find_in_intermediate_node(node[:Kids], key)
             node = node[:Kids][index]
+            node = document.wrap(node, type: self.class) if node
             break unless node && key >= node[:Limits][0] && key <= node[:Limits][1]
           else
             break
@@ -194,7 +195,7 @@ module HexaPDF
         container_name = leaf_node_container_name
         stack = [self]
         until stack.empty?
-          node = stack.pop
+          node = document.wrap(stack.pop, type: self.class)
           if node.key?(container_name)
             data = node[container_name]
             index = 0
@@ -217,7 +218,7 @@ module HexaPDF
       def path_to_key(node, key, stack)
         return unless node.key?(:Kids)
         index = find_in_intermediate_node(node[:Kids], key)
-        stack << node[:Kids][index]
+        stack << document.wrap(node[:Kids][index], type: self.class)
         path_to_key(stack.last, key, stack)
       end
 
@@ -306,6 +307,15 @@ module HexaPDF
       def perform_validation
         super
         container_name = leaf_node_container_name
+
+        if key?(:Kids)
+          self[:Kids].each do |kid|
+            unless kid.indirect?
+              yield("Children of sorted tree nodes must be indirect", true)
+              document.add(kid)
+            end
+          end
+        end
 
         # All keys of the container must be lexically ordered strings and the container must be
         # correctly formatted