+++ CHANGELOG.md
@@ -9,6 +9,8 @@
 
 * Parsing of indirect objects the value of which is an indirect reference
 * Writing of the initial cross-reference section to ensure a single subsection
+* [HexaPDF::Utils::SortedTreeNode] to wrap all /Kids entries with the correct
+  type class
 
 
 ## 1.0.0 - 2024-10-26